localhook x64 bug fixes

+++ src/TestApp/DriverTest.cpp
@@ -59,7 +59,7 @@ void TestDriver()
         {
             PUBLIC_OBJECT_BASIC_INFORMATION info = { 0 };
             ULONG size = 0;
-            status = SAFE_CALL( NtQueryObject, proc.core().handle(), ObjectBasicInformation, &info, sizeof( info ), &size );
+            status = SAFE_CALL( NtQueryObject, proc.core().handle(), ObjectBasicInformation, &info, (ULONG)sizeof( info ), &size );
             if (!NT_SUCCESS( status ))
             {
                 std::cout << "TestDriver: IOCTL_BLACKBONE_GRANT_ACCESS failed, status 0x" << std::hex << status << "\r\n";
